% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reproj-external-types.R, R/reproj.R
\name{reproj.sc}
\alias{reproj.sc}
\alias{reproj.mesh3d}
\alias{reproj.quadmesh}
\alias{reproj.triangmesh}
\alias{reproj}
\alias{reproj.matrix}
\alias{reproj.data.frame}
\alias{reproj_xy}
\alias{reproj_xyz}
\title{Reproject coordinates.}
\usage{
\method{reproj}{sc}(x, target = NULL, ..., source = NULL)

\method{reproj}{mesh3d}(x, target, ..., source = NULL)

\method{reproj}{quadmesh}(x, target, ..., source = NULL)

\method{reproj}{triangmesh}(x, target, ..., source = NULL)

reproj(x, target, ..., source = NULL, four = FALSE)

\method{reproj}{matrix}(x, target, ..., source = NULL, four = FALSE)

\method{reproj}{data.frame}(x, target, ..., source = NULL, four = FALSE)

reproj_xy(x, target, ..., source = NULL)

reproj_xyz(x, target, ..., source = NULL)
}
\arguments{
\item{x}{coordinates}

\item{target}{target specification (PROJ.4 string or epsg code)}

\item{...}{arguments passed to \code{\link[PROJ:proj_trans]{PROJ::proj_trans()}}}

\item{source}{source specification (PROJ.4 string or epsg code)}

\item{four}{defunct, was 'if \code{TRUE} return four columns xyzt', use the shape of the input instead}
}
\value{
numeric matrix of the transformed coordinates, either 2, 3, or 4 columns depending on the
shape of the input. Use \code{\link[=reproj_xy]{reproj_xy()}} or
\code{\link[=reproj_xyz]{reproj_xyz()}} for those specific 2- and 3-column cases.
}
\description{
Reproject coordinates from a matrix or data frame by explicitly specifying the
'source' and 'target' projections.
}
\details{
The transformation engine is the PROJ package, a wrapper around the PROJ
library, so that we can simply give coordinates in data frame
or matrix form, with a source projection and a target projection.

reproj accepts a wide variety of source and target
strings, not just "proj4string", and we are completely subject to the
rules and behaviours of the PROJ library. We always assume "visualization
order", i.e. longitude then latitude, easting then northing (as X, Y).

The basic function \code{\link[=reproj]{reproj()}} takes input in generic form (matrix or data
frame) and returns a matrix with columns matching the coordinate columns
of the input (2, 3, or 4 of x, y, z, m), by
transforming from map projection specified by the  \code{source} argument to that
specified by the \code{target} argument.  Only column order is respected, column
names are ignored.

This model of working also allows adding methods for specific data formats
that already carry a suitable \code{source} projection string. Currently we
support types from the silicate and quadmesh and rgl packages, and only the
\code{target} string need be specified.

This model has obvious flexibility, for packages to import the generic and
call it with the correct \code{source} (from the data format) and the \code{target}
from user, or process controlled mechanism.

The \code{source} argument must be named, and if it is not present a light check
is made that the source data could be "longitude/latitude" and transformation
to \code{target} is applied (this can be controlled by setting options).

Functions \code{\link[=reproj_xy]{reproj_xy()}} and \code{\link[=reproj_xyz]{reproj_xyz()}} are helpers for \code{\link[=reproj]{reproj()}} and always
return 2- or 3-column matrix respectively.

Note that any integer input for \code{source} or \code{target} will be formatted to a
character string like "EPSG:<integer_code>" as a simple convenience. Note that
there are other authorities besides EPSG, so the pattern "AUTH:code" is a general
one and you should really be explicit.

Other R packages for transforming coordinates are geared toward data
that's in a particular format. It's true that only GDAL provides the full
gamut of available geographic map projections, but this leaves a huge variety
of workflows and applications that don't need that level of functionality.
}
\section{Dependencies}{


The \href{https://CRAN.r-project.org/package=PROJ}{PROJ} package (>= 0.7.0)
does all transformation work, wrapping the PROJ library (>= 6.3.1).
}

\section{Global options}{

\subsection{Assuming longitude/latitude input}{

The behaviour is controlled by user-settable options which on start up are
\code{reproj.assume.longlat = TRUE} and
\code{reproj.default.longlat = "OGC:CRS84"}.

If the option \code{reproj.assume.longlat} is set to FALSE then the \code{source}
argument must be named explicitly, i.e. \code{reproj(xy, t_srs, source = s_srs)},
this is to help catch mistakes being made.

If the option \code{reproj.assume.longlat} is set to TRUE and the input data
appear to be sensible longitude/latitude values, then the value of
\code{reproj.default.longlat} is used as the assumed source projection.
}
}

\examples{
reproj(cbind(147, -42), target = "+proj=laea +datum=WGS84",
                         source = getOption("reproj.default.longlat"))
}
