% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reproj-wk.R
\name{reproj.wk_vctr}
\alias{reproj.wk_vctr}
\title{Reproject wk vectors}
\usage{
\method{reproj}{wk_vctr}(x, target, ..., source = NULL)
}
\arguments{
\item{x}{coordinates}

\item{target}{target specification (PROJ.4 string or epsg code)}

\item{...}{arguments passed to \code{\link[proj4:transform]{proj4::ptransform()}}}

\item{source}{source specification (PROJ.4 string or epsg code)}
}
\value{
wk vector
}
\description{
Works for the usual types wkt, wkb, geos, xy, and rct. Beware of blithe reprojection of a
rct, it's not snapping out to the target bounding box (yet).
}
\examples{
if (requireNamespace("PROJ") && requireNamespace("wk")) {
  reproj(wk::xy(147, 0:1, crs = "OGC:CRS84"), "+proj=laea +lon_0=147")
}
}
