% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reproj.R
\name{reproj_xy}
\alias{reproj_xy}
\title{Reproject to specific number of columns}
\usage{
reproj_xy(x, target, ..., source = NULL)
}
\arguments{
\item{x}{coordinates}

\item{target}{target specification (PROJ.4 string or epsg code)}

\item{...}{arguments passed to \code{\link[proj4:transform]{proj4::ptransform()}}}

\item{source}{source specification (PROJ.4 string or epsg code)}
}
\description{
\verb{ _xy}, \verb{_xyz} for the obvious cases.
}
