% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reproj.r
\name{reproj}
\alias{reproj}
\alias{reproj.matrix}
\alias{reproj.data.frame}
\title{Title}
\usage{
reproj(x, source, target, ...)

\method{reproj}{matrix}(x, source, target, ...)

\method{reproj}{data.frame}(x, source, target, ...)
}
\arguments{
\item{x}{coordinates}

\item{source}{source PROJ.4 string}

\item{target}{target PROJ.4 string}

\item{...}{arguments passed to \code{\link{ptransform}}}
}
\value{
matrix
}
\description{
Title
}
